changed config

+++ sft_train.py
@@ -44,7 +44,7 @@ def main(script_args, training_args, model_args):
 
     model = AutoModelForCausalLM.from_pretrained(
         model_args.model_name_or_path,
-        torch_dtype=model_args.dtype,
+        dtype=model_args.dtype,
         use_cache=False if training_args.gradient_checkpointing else True,
         attn_implementation=model_args.attn_implementation,   # <-- ensure it’s used
     )
